rust: Add per-API feature flags
The proxy-api crate always builds servers and clients for all gRPC APIs,
but practically clients usually only intend to pull in a single
client/server for a single API.

This change introduces cargo features which are honored by `build.rs` to
limit which files and functionality are generated.

This change also fixes the `quickcheck::Arbitrary` implementations and
splits them into their respective modules.

CI is updated to exercise individual features in clippy and to use Rust
v1.52.1.

Finally, this change updates all cargo dependencies en masse.
