+++ .github/workflows/ci.yml
@@ -9,9 +9,9 @@ jobs:
   rust_build:
     name: Rust build
     timeout-minutes: 5
-    runs-on: ubuntu-18.04
+    runs-on: ubuntu-latest
     container:
-      image: docker://rust:1.49.0-buster
+      image: docker://rust:1.52.1-buster
     steps:
     # actions/checkout@v2
     - uses: actions/checkout@v2
@@ -39,11 +39,11 @@ jobs:
     timeout-minutes: 5
     runs-on: ubuntu-latest
     container:
-      image: docker://rust:1.49.0-buster
+      image: docker://rust:1.52.1-buster
     steps:
       - uses: actions/checkout@v2
       - run: rustup component add clippy rustfmt
-      - run: cargo clippy --all-targets --all-features
+      - run: make clippy
 
   go_build:
     name: Go build
