+++ rs/Cargo.toml
@@ -7,23 +7,31 @@ publish = false
 edition = "2018"
 
 [features]
+default = []
 # Generate code that is compatible with Tonic's `transport` module.
 transport = ["tonic-build/transport", "tonic/transport"]
 # Generate code that is formatted.
 rustfmt = ["tonic-build/rustfmt"]
 # Enable generation of arbitrary protos with quickcheck.
-arbitrary = ["quickcheck", "rand"]
+arbitrary = ["quickcheck"]
+
+# Determiines whether clients or servers are built
+client = []
+server = []
+
+http_types = ["h2", "http"]
+identity = ["prost-types"]
+destination = ["http_types", "net", "prost-types"]
+net = []
+tap = ["http_types", "net", "prost-types"]
 
 [dependencies]
-tonic = { version = "0.4", default-features = false, features = ["prost", "codegen"] }
+h2 = { version = "0.3", optional = true }
+http = { version = "0.2", optional = true }
 prost = "0.7"
-prost-types = "0.7"
-h2 = "0.3"
-http = "0.2"
-
-# For `arbitrary`:
+prost-types = { version = "0.7", optional = true }
 quickcheck = { version = "1", default-features = false, optional = true }
-rand = { version = "0.8", optional = true }
+tonic = { version = "0.4", default-features = false, features = ["prost", "codegen"] }
 
 [build-dependencies]
 tonic-build = { version = "0.4", default-features = false, features = ["prost"] }
